requirements.txt
httpx>=0.27
google-genai>=0.8
python-telegram-bot[job-queue]>=21.6
python-dotenv>=1.0
pydantic-settings>=2.4
feedparser>=6.0
